:truck: Code Gen: Move files/dirs to make it more clear
