+++ include/sushi/code-generation.h
@@ -1,0 +1,10 @@
+#ifndef SUSHI_ALL_CODE_GENERATION_H
+#define SUSHI_ALL_CODE_GENERATION_H
+
+#include "./code-generation/visitor.h"
+#include "./code-generation/builtin-bash.h"
+#include "./code-generation/code-generation.h"
+#include "./code-generation/scope-manager.h"
+#include "./code-generation/util.h"
+
+#endif
