+++ include/sushi/code-generation/visitor.h
@@ -1,0 +1,10 @@
+#ifndef SUSHI_CODE_GEN_VISITOR_H_
+#define SUSHI_CODE_GEN_VISITOR_H_
+
+#include "./visitor/cmdlike-visitor.h"
+#include "./visitor/expr-visitor.h"
+#include "./visitor/literal-visitor.h"
+#include "./visitor/stmt-visitor.h"
+#include "./visitor/type-visitor.h"
+
+#endif
