+++ src/sushi/code-generation/visitor/cmdlike-visitor.cpp
@@ -1,5 +1,4 @@
-#include "sushi/code-generation/cmdlike-visitor.h"
-#include "sushi/code-generation/util.h"
+#include "sushi/code-generation.h"
 
 namespace sushi {
 namespace code_generation {
